feat: show music island only when player app loses focus
Move the foreign toplevel focus watcher from otto-bar into otto-kit as
a shared utility. otto-islands now tracks which app is focused and
compares it to the MPRIS player name: the music island appears when
the player app loses focus and dismisses when it regains focus.
